ice: implement automatic cleanup of allocations
Use __free() to simplify some paths in the driver and eliminate chances of
missed kfree. Also removes a lot of lines of code.

There are a few more complicated cases where we use no_free_ptr() or
return_ptr() as appropriate, to avoid the auto-free if we want to keep it
on a particular path or success. These are still good, because they
allow us to clean up the goto handlers with the scoped allocations.

Some notes:
Make sure to include the header that implements the kfree DEFINE_FREE().
Always NULL out the pointer when initializing the stack variable to make
sure the cleanup function works correctly for all paths.

Tested locally with KASAN and for basic workloads and driver load/unload
I haven't seen any problems.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>